[WIP] OBJ Loader

+++ bender/bender-base/app/src/main/jni/mesh.h
@@ -31,7 +31,7 @@ public:
   Mesh(Renderer *renderer, std::shared_ptr<Material> material, std::shared_ptr<Geometry> geometryData);
 
   Mesh(Renderer *renderer, std::shared_ptr<Material> material, const std::vector<float>& vertexData,
-          const std::vector<uint16_t>& indexData);
+          const std::vector<uint32_t>& indexData);
 
   Mesh(const Mesh &other, std::shared_ptr<Geometry> geometry);
   Mesh(const Mesh &other, std::shared_ptr<Material> material);
